+++ src/utils/Caret.ts
